fix incorrect mp3 frame size
The mp3 frame size was being calculated incorrectly for MPEG-2.5
streams and multi-channel streams, since the number of channels does
not affect the size of the encoded frame (channels are factored into
the bit rate). These changes correctly calculate the encoded frame
size for multi-channel streams and add support for legacy MPEG-2.5
streams.